+++ packages/staark-common/dst/staark-common.iife.js
@@ -104,10 +104,14 @@
     conditional: () => conditional
   });
   var conditional = (condition, onTruth, onFalse) => {
-    if (condition) {
-      return arrayify(onTruth);
+    let result = condition ? onTruth : onFalse;
+    if (typeof result === "function") {
+      result = result();
     }
-    return arrayify(onFalse ?? []);
+    if (result) {
+      return arrayify(result);
+    }
+    return [];
   };
 
   // src/element.ts
@@ -353,7 +357,14 @@
   });
   var match = (pattern, lookup) => {
     if (lookup && pattern in lookup && lookup[pattern]) {
-      return arrayify(lookup[pattern]);
+      let result = lookup[pattern];
+      if (typeof result === "function") {
+        result = result();
+        if (!result) {
+          return [];
+        }
+      }
+      return arrayify(result);
     }
     return [];
   };