+++ packages/staark-isomorphic/dst/staark-isomorphic.iife.js
@@ -24,10 +24,14 @@
 
   // ../staark-common/src/conditional.ts
   var conditional = (condition, onTruth, onFalse) => {
-    if (condition) {
-      return arrayify(onTruth);
+    let result = condition ? onTruth : onFalse;
+    if (typeof result === "function") {
+      result = result();
     }
-    return arrayify(onFalse ?? []);
+    if (result) {
+      return arrayify(result);
+    }
+    return [];
   };
 
   // ../staark-common/src/marker.ts
@@ -208,7 +212,14 @@
   // ../staark-common/src/match.ts
   var match = (pattern, lookup) => {
     if (lookup && pattern in lookup && lookup[pattern]) {
-      return arrayify(lookup[pattern]);
+      let result = lookup[pattern];
+      if (typeof result === "function") {
+        result = result();
+        if (!result) {
+          return [];
+        }
+      }
+      return arrayify(result);
     }
     return [];
   };