+++ packages/staark-patch/dst/staark-patch.base.iife.js
@@ -257,7 +257,7 @@
                 } else {
                   element2.parentNode.insertBefore(
                     document.createTextNode(childElement),
-                    element2.nextSibling
+                    element2
                   );
                 }
               };