+++ packages/staark/dst/staark.iife.js
@@ -24,10 +24,14 @@
 
   // ../staark-common/src/conditional.ts
   var conditional = (condition, onTruth, onFalse) => {
-    if (condition) {
-      return arrayify(onTruth);
+    let result = condition ? onTruth : onFalse;
+    if (typeof result === "function") {
+      result = result();
     }
-    return arrayify(onFalse ?? []);
+    if (result) {
+      return arrayify(result);
+    }
+    return [];
   };
 
   // ../staark-common/src/marker.ts
@@ -208,7 +212,14 @@
   // ../staark-common/src/match.ts
   var match = (pattern, lookup) => {
     if (lookup && pattern in lookup && lookup[pattern]) {
-      return arrayify(lookup[pattern]);
+      let result = lookup[pattern];
+      if (typeof result === "function") {
+        result = result();
+        if (!result) {
+          return [];
+        }
+      }
+      return arrayify(result);
     }
     return [];
   };
@@ -584,7 +595,7 @@
                 } else {
                   element2.parentNode.insertBefore(
                     document.createTextNode(childElement),
-                    element2.nextSibling
+                    element2
                   );
                 }
               };
@@ -647,9 +658,7 @@
     oldAbstractTree ??= childrenToNodes(_rootElement);
     let active = true, updating = false;
     const updateAbstracts = () => {
-      if (active && !updating && // Only update if changes to the state have been made.
-      proxyChanged && // Don't update while handling listeners.
-      listenerCount <= 0) {
+      if (active && !updating && proxyChanged && listenerCount <= 0) {
         updating = true;
         proxyChanged = false;
         let newAbstractTree = arrayify(